+++ frontend/src/container/App/index.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import FirstPage from "src/component/FirstPage";
 import Header from "src/container/Header";
 import Portfolio from "src/component/Portfolio";
-import Image from "src/container/DisplayImage";
+import Images from "src/container/DisplayImage";
 import { Routes, Route } from "react-router-dom";
 import { MainContext } from "src/context";
 
@@ -36,7 +36,7 @@ function App() {
           <Routes>
             <Route path="*" element={<FirstPage />} />
             <Route path="/portfolio" element={<Portfolio />} />
-            <Route path="/portfolio/:id" element={<Image />} />
+            <Route path="/portfolio/:id" element={<Images />} />
           </Routes>
         )}
       </MainContext.Provider>
