+++ frontend/src/container/DisplayImage/index.tsx
@@ -1,113 +1,34 @@
-import React, { useState, useEffect, useRef, useContext } from "react";
-import { MainContext } from "src/context";
+import React, { useState } from "react";
+import BigImage from "src/container/DisplayImage/Image";
 import { useParams } from "react-router-dom";
-import { DisplayImage } from "src/styled";
-import { Style } from "src/type";
 
-export default function Image({}): JSX.Element {
+const createTab = (length: number): number[] => {
+  let tab = [];
+  for (let i = 0; i < length; i++) {
+    tab.push(i);
+  }
+  return tab;
+};
+
+export default function Images(): JSX.Element {
   const { id } = useParams();
-  const [bigger, setBigger] = useState<number>(0);
-  const ref: any = useRef(null);
-  const { setDisplayImage } = useContext(MainContext);
-  const [onMove, setOnMove] = useState<boolean>(false);
-  let last: number = 0;
-
-  const style: Style[] = [
-    { width: "100%", height: "auto" },
-    { width: "auto", height: "100%" },
-  ];
-
-  useEffect(() => {
-    const handleResize = () => {
-      if (ref.current) {
-        if (ref.current.offsetWidth > window.innerWidth) {
-          setBigger(0);
-        }
-        if (ref.current.offsetHeight > window.innerHeight) {
-          setBigger(1);
-        }
-      }
-    };
-    const handleMove = () => {
-      setOnMove(true);
-      last += 1;
-      const timer = setTimeout(() => {
-        last -= 1;
-        if (last === 0) {
-          setOnMove(false);
-        }
-      }, 1000);
-
-      return () => {
-        clearTimeout(timer);
-      };
-    };
-    window.addEventListener("resize", handleResize);
-    window.addEventListener("mousemove", handleMove);
-    setDisplayImage(true);
-    return () => {
-      window.removeEventListener("resize", handleResize);
-      window.removeEventListener("mousemove", handleResize);
-    };
-  }, []);
-
+  const [display, setDisplay] = useState<number>(parseInt(id || "0"));
   return (
-    <DisplayImage>
-      <div>
-        <div style={{ opacity: onMove ? "1" : "0" }}>
-          <svg
-            aria-label="Fermer"
-            color="#000"
-            fill="#000"
-            height="40"
-            role="img"
-            viewBox="0 0 24 24"
-            width="40"
-          >
-            <polyline
-              fill="none"
-              points="20.643 3.357 12 12 3.353 20.647"
-              stroke="currentColor"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth="3"
-            ></polyline>
-            <line
-              fill="none"
-              stroke="currentColor"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth="3"
-              x1="20.649"
-              x2="3.354"
-              y1="20.649"
-              y2="3.354"
-            ></line>
-          </svg>
-        </div>
-      </div>
-      <img
-        style={style[bigger]}
-        src={require("./landscape.jpg")}
-        alt={`${id}Image`}
-        ref={ref}
-      />
-      <div>
-        <button>
-          <div>
-            <svg width="50px" height="50px" viewBox="0 0 24 24">
-              <path d="M15.41 16.09l-4.58-4.59 4.58-4.59L14 5.5l-6 6 6 6z"></path>
-            </svg>
-          </div>
-        </button>
-        <button>
-          <div>
-            <svg width="50px" height="50px" viewBox="0 0 24 24">
-              <path d="M8.59 16.34l4.58-4.59-4.58-4.59L10 5.75l6 6-6 6z"></path>
-            </svg>
-          </div>
-        </button>
-      </div>
-    </DisplayImage>
+    <div
+      style={{
+        display: "flex",
+        transform: `translateX(${-100 * display}%)`,
+        transition: "transform 0.8s ease-in-out",
+      }}
+    >
+      {createTab(18).map((elt) => (
+        <BigImage
+          key={`${elt}Image`}
+          id={elt}
+          display={display}
+          setDisplay={setDisplay}
+        />
+      ))}
+    </div>
   );
 }
