+++ frontend/src/type.ts
@@ -8,3 +8,9 @@ export type Style = {
   width: string;
   height: string;
 };
+
+export type BigImageProps = {
+  id: number;
+  display: number;
+  setDisplay: React.Dispatch<React.SetStateAction<number>>;
+};
